Refactor: clickHandler를 page단으로 올려서 작성
